+++ Sources/Views/Cells/MessageCollectionViewCell.swift
@@ -38,10 +38,10 @@ open class MessageCollectionViewCell<ContentView: UIView>: UICollectionViewCell,
 
     open var avatarView: AvatarView = AvatarView()
 
-    open var cellTopLabel: MessageLabel = {
-        let topLabel = MessageLabel()
-        topLabel.enabledDetectors = []
-        return topLabel
+    open var cellTopLabel: UILabel = {
+        let label = UILabel()
+        label.numberOfLines = 0
+        return label
     }()
 
     open var messageContentView: ContentView = {
@@ -51,10 +51,10 @@ open class MessageCollectionViewCell<ContentView: UIView>: UICollectionViewCell,
         return contentView
     }()
 
-    open var cellBottomLabel: MessageLabel = {
-        let bottomLabel = MessageLabel()
-        bottomLabel.enabledDetectors = []
-        return bottomLabel
+    open var cellBottomLabel: UILabel = {
+        let label = UILabel()
+        label.numberOfLines = 0
+        return label
     }()
 
     open weak var delegate: MessageCellDelegate?
@@ -77,13 +77,11 @@ open class MessageCollectionViewCell<ContentView: UIView>: UICollectionViewCell,
     // MARK: - Methods
 
     internal func setupSubviews() {
-
         contentView.addSubview(cellTopLabel)
         contentView.addSubview(messageContainerView)
         messageContainerView.addSubview(messageContentView)
         contentView.addSubview(avatarView)
         contentView.addSubview(cellBottomLabel)
-
     }
 
     open override func apply(_ layoutAttributes: UICollectionViewLayoutAttributes) {
@@ -95,12 +93,8 @@ open class MessageCollectionViewCell<ContentView: UIView>: UICollectionViewCell,
 
         messageContainerView.frame = attributes.messageContainerFrame
         messageContentView.frame = messageContainerView.bounds
-
-        cellTopLabel.frame = attributes.cellTopLabelFrame
-        cellTopLabel.textInsets = attributes.cellTopLabelInsets
-
-        cellBottomLabel.frame = attributes.cellBottomLabelFrame
-        cellBottomLabel.textInsets = attributes.cellBottomLabelInsets
+        cellTopLabel.frame = attributes.topLabelFrame
+        cellBottomLabel.frame = attributes.bottomLabelFrame
 
     }